feature/grpc: refactor app
